fix the type for price_amount as it can have decimals

+++ lib/onix/price.rb
@@ -15,7 +15,7 @@ module ONIX
     xml_accessor :class_of_trade, :from => "ClassOfTrade"
     xml_accessor :bic_discount_group_code, :from => "BICDiscountGroupCode"
     xml_accessor :price_status, :from => "PriceStatus", :as => Fixnum, :to_xml => ONIX::Formatters.two_digit
-    xml_accessor :price_amount, :from => "PriceAmount", :as => BigDecimal, :to_xml => ONIX::Formatters.decimal
+    xml_accessor :price_amount, :from => "PriceAmount", :as => Float, :to_xml => ONIX::Formatters.decimal
     xml_accessor :currency_code, :from => "CurrencyCode"
 
     def initialize
